requirements.txt
flask
tensorflow
pillow
numpy
json